Changed to C++20 and warning level 4. Resolved all extra warnings and some general cleanup

+++ VoxelEngine/VoxelEngine/Core/ChunkManager.h
@@ -27,7 +27,7 @@ public:
 	static void UnloadChunk(Chunk* chunk);
 	static void UnloadChunk(const uint32_t& index);
 
-	static const uint16_t GetNumActiveChunks();
+	static const uint32_t GetNumActiveChunks();
 
 	static Chunk* GetChunkAtIndex(const uint16_t index);
